Fix warning "format ‘%u’ expects type ‘unsigned int’, but argument 3 has type ‘size_t’"

+++ src/parsing.c
@@ -451,8 +451,8 @@ int make_default_format_string(scaffolding * dest, unsigned int precision) {
 		const char * const equal_width_base = "%%0%u.%uf";
 		const int left_len = (dest->left < 0) + (size_t)log10(fabs(dest->left)) + 1;
 		const int right_len = (dest->right < 0) + (size_t)log10(fabs(dest->right)) + 1;
-		const size_t pre_dot_digits_wanted = ENUM_MAX(left_len, right_len);
-		const size_t total_chars_wanted = pre_dot_digits_wanted + (precision ? 1 + precision : 0);
+		const unsigned int pre_dot_digits_wanted = ENUM_MAX(left_len, right_len);
+		const unsigned int total_chars_wanted = pre_dot_digits_wanted + (precision ? 1 + precision : 0);
 		const size_t pre_dot_bytes_needed = (size_t)log10(total_chars_wanted) + 1;
 		const size_t base_bytes = strlen(equal_width_base)
 			- strlen("%")
